fix linking paasta_remote_run, drop suppressing/declining offers, deduplicate args parsing
