+++ paasta_tools/cli/cmds/remote_run.py
@@ -15,15 +15,16 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-import sys
-import os
-import shlex
-import service_configuration_lib
 import re
-import uuid
-from datetime import datetime
 
+from paasta_tools.cli.utils import lazy_choices_completer
+from paasta_tools.cli.utils import list_instances
+from paasta_tools.cli.utils import list_clusters
+from paasta_tools.cli.utils import list_services
 from paasta_tools.cli.utils import run_on_master
+from paasta_tools.cli.utils import connectable_master
+from paasta_tools.cli.utils import NoMasterError
+
 from paasta_tools.utils import load_system_paasta_config
 from paasta_tools.utils import paasta_print
 from paasta_tools.utils import PaastaColors
@@ -31,7 +32,58 @@ from paasta_tools.utils import PaastaNotConfiguredError
 from paasta_tools.utils import SystemPaastaConfig
 from paasta_tools.utils import DEFAULT_SOA_DIR
 from paasta_tools.utils import validate_service_instance
-from paasta_tools.paasta_remote_run import add_remote_run_args
+
+
+def add_remote_run_args(parser):
+    parser.add_argument(
+        '-s', '--service',
+        help='The name of the service you wish to inspect',
+    ).completer = lazy_choices_completer(list_services)
+    parser.add_argument(
+        '-c', '--cluster',
+        help=("The name of the cluster you wish to run your task on. "
+              "If omitted, uses the default cluster defined in the paasta remote-run configs"),
+    ).completer = lazy_choices_completer(list_clusters)
+    parser.add_argument(
+        '-y', '--yelpsoa-config-root',
+        dest='yelpsoa_config_root',
+        help='A directory from which yelpsoa-configs should be read from',
+        default=DEFAULT_SOA_DIR,
+    )
+    parser.add_argument(
+        '--json-dict',
+        help='When running dry run, output the arguments as a json dict',
+        action='store_true',
+        dest='dry_run_json_dict',
+    )
+    parser.add_argument(
+        '-C', '--cmd',
+        help=('Run Docker container with particular command, '
+              'for example: "bash". By default will use the command or args specified by the '
+              'soa-configs or what was specified in the Dockerfile'),
+        required=False,
+        default=None,
+    )
+    parser.add_argument(
+        '-i', '--instance',
+        help=("Simulate a docker run for a particular instance of the service, like 'main' or 'canary'"),
+        required=False,
+        default=None,
+    ).completer = lazy_choices_completer(list_instances)
+    parser.add_argument(
+        '-v', '--verbose',
+        help='Show Docker commands output',
+        action='store_true',
+        required=False,
+        default=True,
+    )
+    parser.add_argument(
+        '-d', '--dry-run',
+        help='Don\'t launch the task',
+        action='store_true',
+        required=False,
+        default=False,
+    )
 
 
 def add_subparser(subparsers):
@@ -72,7 +124,7 @@ def paasta_remote_run(args):
         )
         system_paasta_config = SystemPaastaConfig({"volumes": []}, '/etc/paasta')
 
-    cmd_parts = ['/usr/bin/paasta_remote_run.py']
+    cmd_parts = ['/usr/bin/paasta_remote_run']
     args_vars = vars(args)
     args_keys = {
         'service': None,
@@ -102,18 +154,8 @@ def paasta_remote_run(args):
             cmd_parts.extend(['--%s' % arg_key, value])
 
     paasta_print('Running on master: %s' % cmd_parts)
-    if args.very_dry_run:
-        # TODO: maybe print which cluster we'd be connecting to?
-        status = None
-        try:
-            master = connectable_master(args.cluster, system_paasta_config)
-            paasta_print('Very dry run: would have ssh-ed into %s' % master)
-            return_code = 0
-        except NoMasterError as e:
-            paasta_print('Very dry run: could\'t find connectable master %s' % str(e))
-            return_code = err_code
-    else:
-        return_code, status = run_on_master(args.cluster, system_paasta_config, cmd_parts)
+    return_code, status = run_on_master(
+        args.cluster, system_paasta_config, cmd_parts, dry=args.very_dry_run)
 
     # Status results are streamed. This print is for possible error messages.
     if status is not None:
